+++ packages/matter.js-tools/src/testing/web.ts
@@ -17,7 +17,14 @@ import { TestOptions } from "./options.js";
 import { ConsoleProxyReporter, Reporter } from "./reporter.js";
 import type { TestRunner } from "./runner.js";
 
-const libraries = ["chai", "chai-as-promised", "elliptic", "bn.js", "ansi-colors"];
+const libraries = [
+    "chai",
+    "chai-as-promised",
+    "ansi-colors",
+    "@noble/curves/abstract/modular",
+    "@noble/curves/abstract/utils",
+    "@noble/curves/p256",
+];
 
 export async function testWeb(runner: TestRunner, manual: boolean) {
     await buildLibraries();
@@ -123,11 +130,15 @@ function consoleHandler(reporter: Reporter) {
     };
 }
 
+function packageFilename(packageName: string) {
+    return packageName.replace(/^@/, "").replace(/\//g, "_");
+}
+
 function buildIndex(files: string[]) {
     const importMap = JSON.stringify({
         imports: Object.fromEntries(
             libraries.map(name => {
-                let path = `/packages/matter.js-tools/build/lib/${name}`;
+                let path = `/packages/matter.js-tools/build/lib/${packageFilename(name)}`;
                 if (!path.endsWith(".js")) {
                     path = `${path}.js`;
                 }
@@ -160,7 +171,7 @@ function buildIndex(files: string[]) {
 
 async function buildLibraries() {
     for (const name of libraries) {
-        let path = `build/lib/${name}`;
+        let path = `build/lib/${packageFilename(name)}`;
         if (!path.endsWith(".js")) {
             path = `${path}.js`;
         }
@@ -171,10 +182,13 @@ async function buildLibraries() {
         }
 
         await build({
-            entryPoints: [Package.workspace.resolve(`node_modules/${name}`)],
+            entryPoints: [Package.workspace.findExport(name)],
             bundle: true,
             format: "esm",
             outfile,
+
+            // Grr esbuild doesn't respect this
+            //logOverride: { "empty-import-meta": "silent" },
         });
     }
 }